# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Educativo Virtual - API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/acad/buzon-sugerencias": {
            "get": {
                "tags": [
                    "Buzón de sugerencias"
                ],
                "summary": "Permite a un estudiante obtener la lista de sus sugerencias registradas.",
                "operationId": "f5bace4d0390a02a94f1d394b5846438",
                "responses": {
                    "200": {
                        "description": "Sugerencia registrada exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Buzón de sugerencias"
                ],
                "summary": "Permite a un estudiante registrar una sugerencia en el buzón de sugerencias.",
                "operationId": "ce6b70cab7f63c3c588da0f6afb4dc7f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/iCredEntPerfId"
                    },
                    {
                        "name": "cAsunto",
                        "in": "query",
                        "description": "Asunto de la sugerencia",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "cSugerencia",
                        "in": "query",
                        "description": "Contenido de la sugerencia",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "iPrioridadId",
                        "in": "query",
                        "description": "ID de la prioridad de la sugerencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Sugerencia registrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Se ha registrado su sugerencia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Problema al registrar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "danger"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Hubo un problema al registrar: [detalle del error]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Login de usuario",
                "description": "Autentica un usuario y retorna un token JWT junto con la data del usuario.",
                "operationId": "d0fabbdd66ee2d9f03f4ab51b6928b21",
                "parameters": [
                    {
                        "name": "user",
                        "in": "query",
                        "description": "Usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pass",
                        "in": "query",
                        "description": "Contraseña del usuario",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 6
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Errores de autenticación posibles:\n     *                      - Ya alcanzó el límite de intentos\n     *                      - Usuario y contraseña incorrectos\n     *                      - Debe actualizar su contraseña\n     *                      - Usuario no existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "validated": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Verifica tu usuario y contraseña"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Inicio de sesión correcto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "accessToken": {
                                            "description": "JWT token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "description": "Tiempo de duración del token",
                                            "type": "integer"
                                        },
                                        "user": {
                                            "description": "Data del usuario",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "parameters": {
            "iCredEntPerfId": {
                "name": "iCredEntPerfId",
                "in": "header",
                "description": "Id de la credencial a nivel entidad y perfil",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "Accept": {
                "name": "Accept",
                "in": "header",
                "description": "Tipo de respuesta esperada",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": "application/json"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Buzón de sugerencias",
            "description": "Operaciones relacionadas con el registro de sugerencias por parte de estudiantes"
        },
        {
            "name": "Autenticación",
            "description": "Inicio de sesión de usuarios"
        }
    ],
    "security": {
        "passport": {
            "type": "oauth2",
            "description": "Laravel passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "http://localhost/oauth/authorize",
                    "tokenUrl": "http://localhost/oauth/token",
                    "refreshUrl": "http://localhost/token/refresh",
                    "scopes": []
                }
            }
        }
    }
}